Reservation Delete, GetAllReservation Methods

+++ src/main/java/com/saferent/service/CarService.java
@@ -24,11 +24,14 @@ public class CarService {
 
     private final ImageFileService imageFileService;
 
+    private final ReservationService reservationService;
+
     private final CarMapper carMapper;
 
-    public CarService(CarRepository carRepository, ImageFileService imageFileService, CarMapper carMapper) {
+    public CarService(CarRepository carRepository, ImageFileService imageFileService, ReservationService reservationService, CarMapper carMapper) {
         this.carRepository = carRepository;
         this.imageFileService = imageFileService;
+        this.reservationService = reservationService;
         this.carMapper = carMapper;
     }
 
@@ -132,6 +135,12 @@ public class CarService {
 
         if (car.getBuiltIn()) {
             throw new BadRequestException(ErrorMessage.NOT_PERMITTED_METHOD_MESSAGE);
+        }
+
+        boolean exist = reservationService.existByCar(car);
+
+        if (exist) {
+            throw new BadRequestException(ErrorMessage.CAR_USED_BY_RESERVATION_MESSAGE);
         }
 
         carRepository.delete(car);
